refactor: remove some depinject usage
