+++ client/tx/aux_builder_test.go
@@ -5,29 +5,23 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	_ "cosmossdk.io/api/cosmos/bank/v1beta1"
-	"cosmossdk.io/depinject"
-	clienttestutil "github.com/cosmos/cosmos-sdk/client/testutil"
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
 	typestx "github.com/cosmos/cosmos-sdk/types/tx"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	"github.com/cosmos/cosmos-sdk/x/bank"
 )
 
 func TestAuxTxBuilder(t *testing.T) {
-	var (
-		reg codectypes.InterfaceRegistry
-		cdc codec.Codec
-	)
-	err := depinject.Inject(clienttestutil.TestConfig, &reg, &cdc)
 	bankModule := bank.AppModuleBasic{}
+	cdc := moduletestutil.MakeTestEncodingConfig(bankModule).Codec
+	reg := codectypes.NewInterfaceRegistry()
 
-	require.NoError(t, err)
 	testdata.RegisterInterfaces(reg)
 	// required for test case: "GetAuxSignerData works for DIRECT_AUX"
 	bankModule.RegisterInterfaces(reg)
