--- x/params/testutil/app_config.go
@@ -1,23 +1,0 @@
-package testutil
-
-import (
-	"github.com/cosmos/cosmos-sdk/testutil/configurator"
-
-	_ "github.com/cosmos/cosmos-sdk/x/auth"           // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/auth/tx/config" // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/bank"           // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/consensus"      // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/genutil"        // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/params"         // import as blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/staking"        // import as blank for app wiring
-)
-
-var AppConfig = configurator.NewAppConfig(
-	configurator.AuthModule(),
-	configurator.BankModule(),
-	configurator.StakingModule(),
-	configurator.TxModule(),
-	configurator.ConsensusModule(),
-	configurator.ParamsModule(),
-	configurator.GenutilModule(),
-)
